requirements.txt
pynput
pygame
pystray
Pillow
tkinter